rename type file and set selectors
