+++ src/App.css
@@ -10,3 +10,7 @@
 * {
   font-family: var(--font-family);
 }
+
+body::-webkit-scrollbar {
+  display: none;
+}
