+++ src/components/common-nav/CommonNav.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Col, Row, Image } from "antd";
-import { CommonNav } from "@/types/HeaderTypes";
+import { CommonNav } from "@/types/GenericTypes";
 
 const CommonNavComponent: React.FC<CommonNav> = ({
   logo,
